Disable the loadLogs() call in the delete event listener.
It was likely to cause a double reload since the log list
is already refreshed on window focus. The SDK table row
delete method hides the row anyway, and nothing is using the
table data array index, so an out-of-sync array shouldn't
cause a problem...as the row generates it's own detail
view when it is selected.