Moved log list activity indicator to nav bar
Activity indicator on the nav bar for iphone only. Kept the
indicator as a window overlay for android.

Also, removed the call to refresh the log list after deleting a log
in the detail page, since the list is refreshed when the table
view regains focus anyway. This is still not ideal, since it causes
a signifigant delay (5-10s) when there are a lot (>100) of logs.

Really, want to only refresh the log list when there are new log files,
eg. when logging was started on the dashboard, but after the log list
view has been opened (and generated) for the first time. Need a way to
'dirty' the log list, triggering an update when necessary.